All entity identifiers now use a uuid string.
The entities have been annotated with the following hibernate provided generator
  @GeneratedValue(generator = "system-uuid")
  @GenericGenerator(name = "system-uuid", strategy = "uuid.hex")

This changes the storage datatype of the id column for all the entities.

Certificate serial numbers will remain numbers so that they are compliant with
the x509 specification.

+++ proxy/src/main/java/org/fedoraproject/candlepin/model/EntitlementCurator.java
@@ -14,7 +14,6 @@
  */
 package org.fedoraproject.candlepin.model;
 
-import java.math.BigInteger;
 import java.util.HashSet;
 import java.util.LinkedList;
 import java.util.List;
@@ -107,10 +106,10 @@ public class EntitlementCurator extends AbstractHibernateCurator<Entitlement> {
     }
     
     @Transactional
-    public Entitlement findByCertificateSerial(BigInteger serial) {
+    public Entitlement findByCertificateSerial(Long serial) {
         return (Entitlement) currentSession().createCriteria(Entitlement.class)
             .createCriteria("certificates")
-                .add(Restrictions.eq("serial.id", serial.longValue()))
+                .add(Restrictions.eq("serial.id", serial))
             .uniqueResult();
     }
